# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: RMS Implementation Architecture Baseline 2.0 — OpenAPI draft
  version: 2.0.0-draft
  description: Design only; NOT IMPLEMENTED. 业务来源唯一冻结Baseline1.3。服务器根/api/v1。全部保护操作检查session、启用状态、实际角色并集；自我评审等限制独立执行。请求只白名单，历史不可变，无generic
    status API。SQL BIGINT IDs/lockVersion采用十进制字符串。两个/auth技术GET为实现派生，不新增FR。数据库generated活动键不公开。密码原值与1..72 UTF-8 bytes技术政策已冻结；非admin assignee目录待决。
servers:
- url: /api/v1
security:
- cookieSession: []
...
paths:
...
  /relations:
    post:
      operationId: createRelation
      tags:
      - relation
      summary: 图锁和同连接事务中校验两端、归一化对称关系、拒绝重复和DEPENDS_ON环
      description: 图锁和同连接事务中校验两端、归一化对称关系、拒绝重复和DEPENDS_ON环。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: relation
      x-fr:
      - FR-AUDIT-001
      - FR-REL-001
      - FR-REL-003
      - FR-REL-005
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REL-001
      - BR-REL-002
      - BR-REL-003
      - BR-REL-004
      - BR-REL-005
      - BR-REL-006
      - BR-REL-007
      - BR-REL-010
      x-uc:
      - UC-07
      - UC-S1
      x-ac:
      - AC-07
      - AC-10
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '201':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementRelationResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRelationRequest'
  /relations/{id}:
    delete:
      operationId: deleteRelation
      tags:
      - relation
      summary: 同图锁解除关系，保存前值审计，不删除两端需求
      description: 同图锁解除关系，保存前值审计，不删除两端需求。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: relation
      x-fr:
      - FR-AUDIT-001
      - FR-REL-002
      - FR-REL-005
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REL-001
      - BR-REL-005
      - BR-REL-006
      - BR-REL-007
      - BR-REL-008
      - BR-REL-009
      - BR-REL-010
      x-uc:
      - UC-07
      - UC-S1
      x-ac:
      - AC-07
      - AC-10
      x-required-roles: *id003
      x-transaction: APPLICATION_REQUIRED
      responses:
        '204':
          description: 成功；具体业务含义见操作说明
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: expectedSourceLockVersion
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/LockVersion'
        description: expectedSourceLockVersion
      - name: expectedTargetLockVersion
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/LockVersion'
        description: expectedTargetLockVersion
      - $ref: '#/components/parameters/CsrfHeader'
  /requirements/{id}/relations:
    get:
      operationId: listRelations
      tags:
      - relation
      summary: 读取当前入向/出向关系与类型
      description: 读取当前入向/出向关系与类型。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: relation
      x-fr:
      - FR-REL-004
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REL-002
      - BR-REL-003
      - BR-REL-004
      x-uc:
      - UC-07
      x-ac:
      - AC-07
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementRelationResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
  /requirements/{id}/trace:
    get:
      operationId: traceRequirements
      tags:
      - relation
      summary: 有限深度遍历当前图，visited去重，保留方向和类型
      description: 有限深度遍历当前图，visited去重，保留方向和类型。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: relation
      x-fr:
      - FR-REL-006
      x-br:
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-REL-003
      - BR-REL-004
      - BR-REL-007
      x-uc:
      - UC-07
      x-ac:
      - AC-07
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: depth
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 10
          default: 2
        description: 实现安全边界，有限深度；不改变关系语义
      - name: relationType
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/RelationType'
        description: relationType
...
  /requirements/{id}/comments:
    get:
      operationId: listComments
      tags:
      - comment
      summary: 读取评论安全投影，已删正文不返回
      description: 读取评论安全投影，已删正文不返回。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: comment
      x-fr:
      - FR-COM-002
      - FR-COM-003
      x-br:
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-COM-003
      - BR-COM-004
      - BR-COM-005
      - BR-REQ-012
      x-uc:
      - UC-09
      x-ac:
      - AC-08
      x-required-roles: *id001
      x-transaction: READ_ONLY
      responses:
        '200':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponsePage'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
    post:
      operationId: createComment
      tags:
      - comment
      summary: 未撤销需求发布非空评论，作者和时间由服务端生成
      description: 未撤销需求发布非空评论，作者和时间由服务端生成。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: comment
      x-fr:
      - FR-AUDIT-001
      - FR-COM-001
      - FR-COM-003
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-COM-001
      - BR-COM-002
      - BR-COM-003
      - BR-COM-004
      - BR-COM-006
      x-uc:
      - UC-09
      - UC-S1
      x-ac:
      - AC-08
      - AC-10
      x-required-roles:
      - REQUIREMENT_ENGINEER
      - REVIEWER
      - PROJECT_MEMBER
      x-transaction: APPLICATION_REQUIRED
      responses:
        '201':
          description: 成功；具体业务含义见操作说明
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
  /comments/{id}:
    delete:
      operationId: deleteOwnComment
      tags:
      - comment
      summary: 仅作者本人逻辑删除，数据库保留正文/作者/时间及删除证据
      description: 仅作者本人逻辑删除，数据库保留正文/作者/时间及删除证据。仅启用账号；服务端重新检查角色并集及独立业务限制。
      x-module: comment
      x-fr:
      - FR-AUDIT-001
      - FR-COM-004
      x-br:
      - BR-AUDIT-001
      - BR-AUDIT-002
      - BR-AUDIT-003
      - BR-AUDIT-004
      - BR-AUDIT-006
      - BR-AUTH-001
      - BR-AUTH-002
      - BR-AUTH-003
      - BR-AUTH-004
      - BR-COM-002
      - BR-COM-005
      x-uc:
      - UC-09
      - UC-S1
      x-ac:
      - AC-08
      - AC-10
      x-required-roles:
      - REQUIREMENT_ENGINEER
      - REVIEWER
      - PROJECT_MEMBER
      x-transaction: APPLICATION_REQUIRED
      responses:
        '204':
          description: 成功；具体业务含义见操作说明
        '400':
          description: 字段/完整性错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证/凭证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限/启用状态/自我评审/CSRF拒绝
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 修订/状态/唯一/基准/图锁冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: 内部错误；业务写全回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/CsrfHeader'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: HttpOnly Secure SameSite=Lax，同源session；所有unsafe请求另需X-CSRF-TOKEN，包括login/logout。
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: GET /auth/csrf取得；登录/退出后刷新；不是actor身份。
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      maxLength: 19
      description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
    LockVersion:
      type: string
      pattern: ^(0|[1-9][0-9]*)$
      maxLength: 19
      description: 原有signed BIGINT lockVersion非负十进制字符串；服务端校验<=9223372036854775807、比较并由@Version更新。
...
    RelationType:
      type: string
      enum:
      - DEPENDS_ON
      - REFINES
      - DERIVED_FROM
      - CONFLICTS_WITH
      - DUPLICATES
      - RELATES_TO
    RequirementResponse:
      type: object
      additionalProperties: false
      properties:
        requirementId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        requirementKey:
          type: string
          maxLength: 32
          readOnly: true
        title:
          type: string
          maxLength: 200
          readOnly: true
        description:
          type: string
          readOnly: true
        level:
          type: string
          maxLength: 32
          enum:
          - BUSINESS
          - USER
          - SYSTEM
          readOnly: true
        kind:
          type: string
          maxLength: 32
          enum:
          - FUNCTIONAL
          - QUALITY
          - CONSTRAINT
          readOnly: true
        priority:
          type: string
          maxLength: 32
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
          readOnly: true
        source:
          type:
          - string
          - 'null'
          readOnly: true
        rationale:
          type:
          - string
          - 'null'
          readOnly: true
        acceptanceCriteria:
          type:
          - string
          - 'null'
          readOnly: true
        status:
          type: string
          maxLength: 32
          enum:
          - DRAFT
          - UNDER_REVIEW
          - APPROVED
          - REJECTED
          - IMPLEMENTED
          - VERIFIED
          readOnly: true
        creatorId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        assigneeId:
          type:
          - string
          - 'null'
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        currentVersionId:
          type:
          - string
          - 'null'
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        firstSubmittedAt:
          type:
          - string
          - 'null'
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        isWithdrawn:
          type: integer
          minimum: 0
          enum:
          - 0
          - 1
          readOnly: true
        withdrawnBy:
          type:
          - string
          - 'null'
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        withdrawnAt:
          type:
          - string
          - 'null'
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        createdAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        lockVersion:
          type: string
          pattern: ^(0|[1-9][0-9]*)$
          maxLength: 19
          description: 原有signed BIGINT lockVersion非负十进制字符串；服务端校验<=9223372036854775807、比较并由@Version更新。
          readOnly: true
        creator:
          $ref: '#/components/schemas/UserSummary'
        assignee:
          oneOf:
          - $ref: '#/components/schemas/UserSummary'
          - type: 'null'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
      required:
      - requirementId
      - requirementKey
      - title
      - description
      - level
      - kind
      - priority
      - source
      - rationale
      - acceptanceCriteria
      - status
      - creatorId
      - assigneeId
      - currentVersionId
      - firstSubmittedAt
      - isWithdrawn
      - withdrawnBy
      - withdrawnAt
      - createdAt
      - updatedAt
      - lockVersion
      - creator
      - assignee
      - tags
      description: 安全传输投影；不是JPA Entity；nullable逐项沿用Baseline 1.3，数据库generated活动键不公开。
...
    RequirementRelationResponse:
      type: object
      additionalProperties: false
      properties:
        relationId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        sourceRequirementId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        targetRequirementId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        relationType:
          type: string
          maxLength: 32
          enum:
          - DEPENDS_ON
          - REFINES
          - DERIVED_FROM
          - CONFLICTS_WITH
          - DUPLICATES
          - RELATES_TO
          readOnly: true
        description:
          type:
          - string
          - 'null'
          maxLength: 1000
          readOnly: true
        createdBy:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        createdAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
      required:
      - relationId
      - sourceRequirementId
      - targetRequirementId
      - relationType
      - description
      - createdBy
      - createdAt
      description: 安全传输投影；不是JPA Entity；nullable逐项沿用Baseline 1.3，数据库generated活动键不公开。
...
    UserSummary:
      type: object
      additionalProperties: false
      properties:
        userId:
          $ref: '#/components/schemas/Id'
        displayName:
          type: string
          maxLength: 100
        accountStatus:
          type: string
          enum:
          - ENABLED
          - DISABLED
      required:
      - userId
      - displayName
      - accountStatus
...
    CommentResponse:
      type: object
      additionalProperties: false
      properties:
        commentId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        requirementId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        authorId:
          type: string
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        content:
          type:
          - string
          - 'null'
          description: isDeleted=1时必须为null，不返回已删除正文；原数据库content仍NOT NULL并保留。
        createdAt:
          type: string
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
        isDeleted:
          type: integer
          minimum: 0
          enum:
          - 0
          - 1
          readOnly: true
        deletedBy:
          type:
          - string
          - 'null'
          pattern: ^[1-9][0-9]*$
          maxLength: 19
          description: SQL signed BIGINT正整数十进制字符串；服务端再校验<=9223372036854775807，避免JavaScript Number精度丢失。
          readOnly: true
        deletedAt:
          type:
          - string
          - 'null'
          format: date-time
          pattern: .*Z$
          description: 服务端UTC，ISO8601 Z；SQL DATETIME(6)，微秒精度。
          readOnly: true
      required:
      - commentId
      - requirementId
      - authorId
      - content
      - createdAt
      - isDeleted
      - deletedBy
      - deletedAt
...
    CreateRelationRequest:
      type: object
      additionalProperties: false
      properties:
        sourceRequirementId:
          $ref: '#/components/schemas/Id'
        targetRequirementId:
          $ref: '#/components/schemas/Id'
        relationType:
          $ref: '#/components/schemas/RelationType'
        description:
          type:
          - string
          - 'null'
          maxLength: 1000
        expectedSourceLockVersion:
          $ref: '#/components/schemas/LockVersion'
        expectedTargetLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - sourceRequirementId
      - targetRequirementId
      - relationType
      - expectedSourceLockVersion
      - expectedTargetLockVersion
    CreateCommentRequest:
      type: object
      additionalProperties: false
      properties:
        content:
          type: string
          minLength: 1
          pattern: .*\S.*
      required:
      - content
...
    TraceResponse:
      type: object
      additionalProperties: false
      properties:
        rootRequirementId:
          $ref: '#/components/schemas/Id'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/RequirementResponse'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/RequirementRelationResponse'
        depth:
          type: integer
          minimum: 1
        truncated:
          type: boolean
      required:
      - rootRequirementId
      - nodes
      - edges
      - depth
      - truncated
      description: 当前关系图；按visited去重，方向及类型保留；depth有限，包含被截断标识。
    ApiError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
          - INVALID_INPUT
          - INCOMPLETE_CONTENT
          - UNAUTHENTICATED
          - INVALID_CREDENTIALS
          - FORBIDDEN
          - ACCOUNT_DISABLED
          - SELF_REVIEW
          - NOT_AUTHOR
          - NOT_FOUND
          - LOCK_VERSION_CONFLICT
          - STATE_CONFLICT
          - BASE_VERSION_STALE
          - DUPLICATE
          - ACTIVE_CHANGE_EXISTS
          - ACTIVE_REVIEW_EXISTS
          - DEPENDENCY_CYCLE
          - GRAPH_BUSY
          - INTERNAL_ERROR
        message:
          type: string
        correlationId:
          type: string
        currentLockVersion:
          $ref: '#/components/schemas/LockVersion'
      required:
      - code
      - message
      - correlationId
      description: 安全错误，无SQL/stack/密码；currentLockVersion仅有明确受权资源的修订冲突才提供。
...
    RequirementRelationResponsePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequirementRelationResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - size
      - totalElements
...
    CommentResponsePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
      required:
      - items
      - page
      - size
      - totalElements
...